requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.4
requests==2.31.0
joblib==1.3.2
scikit-learn==1.3.2
pytz==2024.1
